# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EarScope API",
        "description": "Dokumentasi API untuk EarScope",
        "version": "1.0"
    },
    "paths": {
...
        "/api/penanganan": {
            "get": {
                "tags": [
                    "Penanganan"
                ],
                "summary": "Mendapatkan daftar penanganan",
                "description": "Dokter dapat melihat semua penanganan, Pasien hanya melihat penanganan miliknya",
                "operationId": "28f0e17a38f73aa3af72d86c18bed6a0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "selesai",
                                "dibatalkan"
                            ]
                        }
                    },
                    {
                        "name": "tanggal_mulai",
                        "in": "query",
                        "description": "Filter tanggal mulai (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "tanggal_akhir",
                        "in": "query",
                        "description": "Filter tanggal akhir (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data penanganan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data penanganan berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PenangananResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Penanganan"
                ],
                "summary": "Membuat penanganan baru",
                "description": "Hanya dokter yang dapat membuat penanganan baru",
                "operationId": "f91cae4ee2f5d6eb1deda846865bf5c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "tanggal_penanganan",
                                    "keluhan",
                                    "telinga_terkena"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID pasien",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tanggal_penanganan": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-15"
                                    },
                                    "keluhan": {
                                        "type": "string",
                                        "example": "Telinga terasa sakit dan berdengung"
                                    },
                                    "riwayat_penyakit": {
                                        "type": "string",
                                        "example": "Pernah mengalami infeksi telinga"
                                    },
                                    "diagnosis_manual": {
                                        "type": "string",
                                        "example": "Otitis media akut"
                                    },
                                    "telinga_terkena": {
                                        "type": "string",
                                        "enum": [
                                            "kiri",
                                            "kanan",
                                            "keduanya"
                                        ],
                                        "example": "kiri"
                                    },
                                    "tindakan": {
                                        "type": "string",
                                        "example": "Pemberian antibiotik dan tetes telinga"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "selesai",
                                            "dibatalkan"
                                        ],
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Penanganan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Penanganan berhasil dibuat"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PenangananResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penanganan/{id}": {
            "get": {
                "tags": [
                    "Penanganan"
                ],
                "summary": "Mendapatkan detail penanganan",
                "operationId": "d1249e57505612ad68dc24a8a2a11800",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail penanganan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail penanganan berhasil diambil"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PenangananResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Penanganan"
                ],
                "summary": "Mengupdate penanganan",
                "description": "Hanya dokter yang dapat mengupdate penanganan",
                "operationId": "b88240be9fcc2ea428ad9ce6b0761409",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tanggal_penanganan": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "keluhan": {
                                        "type": "string"
                                    },
                                    "riwayat_penyakit": {
                                        "type": "string"
                                    },
                                    "diagnosis_manual": {
                                        "type": "string"
                                    },
                                    "telinga_terkena": {
                                        "type": "string",
                                        "enum": [
                                            "kiri",
                                            "kanan",
                                            "keduanya"
                                        ]
                                    },
                                    "tindakan": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "selesai",
                                            "dibatalkan"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Penanganan berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Penanganan"
                ],
                "summary": "Menghapus penanganan",
                "description": "Hanya dokter yang dapat menghapus penanganan",
                "operationId": "b52d1257d63f9e689166ae63d890532a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penanganan berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penanganan/statistik": {
            "get": {
                "tags": [
                    "Penanganan"
                ],
                "summary": "Mendapatkan statistik penanganan",
                "description": "Hanya dokter yang dapat melihat statistik",
                "operationId": "99d30b520491bd50b6477990a5a83822",
                "responses": {
                    "200": {
                        "description": "Statistik berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token di kolom ini dengan format Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}